client: split from main.rs
