+++ src/commands/moderation/clear_messages.rs
@@ -1,4 +1,4 @@
-use crate::{Context, Error};
+use crate::client::{Context, Error};
 
 use log::debug;
 use poise::serenity_prelude::GetMessages;
