+++ src/handlers/event/message.rs
@@ -1,4 +1,4 @@
-use crate::{consts, Data};
+use crate::{client::Data, consts};
 
 use eyre::{eyre, Result};
 use log::{debug, warn};
